requirements.txt
pbr>=0.6,!=0.7,<1.0
oslo.config
prettytable
requests
pyzabbix
PyYaml
